Moved price field to mixin trait.
