+++ src/test/scala-2.11/markets/orders/TestOrder.scala
@@ -17,12 +17,7 @@ package markets.orders
 
 import java.util.UUID
 
-import markets.tradables.{Security, Tradable}
+import markets.tradables.Tradable
 
 
-case class TestOrder(issuer: UUID,
-                     price: Long,
-                     quantity: Long,
-                     timestamp: Long,
-                     tradable: Tradable,
-                     uuid: UUID) extends Order
+case class TestOrder(issuer: UUID, quantity: Long, timestamp: Long, tradable: Tradable, uuid: UUID) extends Order
